+++ CefSharp.BrowserSubprocess.Core/CefSharp.BrowserSubprocess.Core.vcxproj.filters
@@ -113,6 +113,9 @@
     <ClInclude Include="SubProcessApp.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="BindObjectAsyncHandler.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="AssemblyInfo.cpp">
